requirements.txt
# --- FastAPI Core ---
fastapi==0.104.1
uvicorn[standard]==0.24.0

# --- Database and ORM ---
sqlmodel==0.0.16
pymysql==1.1.0
alembic==1.13.1

# --- Auth, Security, Env ---
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.7

# --- Security Hardening Dependencies ---
bleach==6.2.0  # Input sanitization
python-magic==0.4.27  # File type detection for validation
types-requests==2.32.0.20250602  # Type stubs for requests
types-redis==4.6.0.20250602  # Type stubs for redis
types-python-jose==3.5.0.20250531  # Type stubs for python-jose
types-passlib==1.7.7.20250602  # Type stubs for passlib
pandas-stubs==2.2.3.250527  # Type stubs for pandas
types-openpyxl==3.1.5.20250602  # Type stubs for openpyxl
types-fpdf2==2.8.3.20250516  # Type stubs for fpdf2
types-bleach==6.2.0.20250514  # Type stubs for bleach
types-html5lib==1.1.11.20250516  # Type stubs for html5lib (bleach dependency)

# --- Data Handling and Parsing ---
pandas==2.3.0
openpyxl==3.1.5
python-docx==1.2.0
pypdfium2==4.30.1
pdfplumber==0.11.4

# --- PDF Generation ---
fpdf2==2.8.1

# --- AI/NLP Integration ---
google-generativeai==0.8.3
openai==1.30.1
transformers==4.52.4
huggingface-hub==0.33.0

# --- Email Support ---
jinja2==3.1.6
msal==1.30.0
requests==2.32.0

# --- Async HTTP/Utils ---
httpx==0.27.0

# --- Caching and Performance ---
redis==5.0.1  # Optional: for advanced caching
aiofiles==23.2.1  # For async file operations

# --- Monitoring and Observability ---
prometheus-client==0.19.0  # For metrics export
structlog==23.2.0  # Structured logging

# --- Testing ---
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==4.0.0

# --- Linting, Formatting, Dev Tools ---
black==24.4.2
flake8==7.0.0
isort==5.13.2
mypy==1.10.0
pre-commit==3.7.1

# --- Other dependencies for parsing ---
pdfminer.six==20250506
python-dateutil==2.9.0.post0

# --- Data Science/Analytics ---
numpy==2.3.0
scikit-learn==1.7.0
scipy==1.15.3
matplotlib==3.8.2  # For chart generation
seaborn==0.13.2  # Enhanced plotting

# --- Utilities ---
tqdm==4.67.1
click==8.1.7  # For CLI tools
pydantic-settings==2.1.0  # Enhanced settings management

# --- Security Dependencies ---
python-magic==0.4.27  # For MIME type detection
slowapi==0.1.9  # Rate limiting middleware (Redis-backed)
boto3==1.34.144  # AWS SDK for Secrets Manager
azure-keyvault-secrets==4.8.0  # Azure Key Vault
azure-identity==1.17.1  # Azure authentication

# --- Additional Security ---
cryptography>=41.0.7  # Enhanced crypto support
pycryptodome==3.20.0  # Additional crypto utilities
